Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/runtime-error-in-body-onerror-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: undefined_variable
CONSOLE MESSAGE: ReferenceError: Can't find variable: undefined_variable_in_onerror

PASS runtime error in <body onerror>
